+++ __tests__/resources/nested-other-other.js
@@ -1,0 +1,11 @@
+#!/usr/bin/env node
+
+"use strict";
+
+const path = require("path");
+
+const dir = path.join(__dirname, "../other-other-dir");
+
+setTimeout(() => {
+  process.stdout.write(dir);
+}, 1500);